+++ packages/observable-client/src/chainHead/streams/get-runtime-creator.ts
@@ -104,7 +104,7 @@ const withRecovery =
   }
 
 export const getRuntimeCreator = (
-  call$: (hash: string, method: string, args: string) => Observable<string>,
+  call$: (hash: string, method: string, args: string) => Observable<Uint8Array>,
   getCachedMetadata: (codeHash: string) => Observable<Uint8Array | null>,
   setCachedMetadata: (codeHash: string, metadataRaw: Uint8Array) => void,
 ) => {